django: move django* packages under django submenu
This changes the python[3]-django dependencies in packages to be
non-selecting, and adds an MDEPENDS line so that the *-src packages get
placed inside the django menu as well.

Added MENU:= to the src-package definitions in python[3]-package.mk,
so it does not import that setting from the binary package.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>