seafile-seahub: use python-pillow as dep
The pillow package has been updated to the Python[3] packaging format, and
now the package names are `python-pillow` & `python3-pillow`.
This change updates seafile-seahub to use it.

Not updating other packages as they will be converted to Python[3]
packaging format.

And not bumping PKG_RELEASE here as it will be done in the last commit that
updates deps for seafile-seahub.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>