luasocket: update version
I made the following modifications:

Makefile - Updated the upstream version. Set the LUAV environment
variable in the build variant. Set an include path in the lua build so
that that lua headers could be found.  Set source date version to
identify the source.

0001-Add-interface-support.patch - ported this patch which included
replacing one deprecated call for 5.3. Validated that the interface
functionality added by the patch is working and functional. The
following lua script line was functional: print(socket.iface.indextoname(2))

0301-Fix-mpc85xx-build.patch - Just fixed the offsets and fuzz in the
patch

040-remove-fpic-and-warnings.patch - ported this patch. I did make one
change. The patch name is called remove fpic and warnings, but the
patch was also removing optimization. I ported the removal of fpic and
warnings, however I left optimization in.

Built and tested on BCM5301X.

Signed-off-by: Colby Whitney <colby.whitney@luxul.com>
[ rmilecki: fix whitespaces in 040-remove-fpic-and-warnings.patch ]
Signed-off-by: Rafał Miłecki <rafal@milecki.pl>