boost: Fix fiber compilation for mips
The issue was that the pause instruction was emitted through an asm
directive which was not correct for some mips platforms.

Simplified boost-fiber-exclude as a result.

Removed uClibc-ng math patch. It was not correct as it broke float and
long double support (std variants use function overloads). A different
solution was applied upstream. As it's quite annoying to backport, just
wait until a new release comes with that change. ARC as a platform is
barely supported anyways.

Swapped asio patch for the upstream submission, which is unfortunately, in
limbo.

Refreshed remaining patch.

Signed-off-by: Rosen Penev <rosenp@gmail.com>