libgd: fix version in generated pkg-config .pc file
With the previous change to cmake, the generated pkg-config .pc
file does not contain the library version anymore. This breaks
programs which checks for a specific version, e.g. upcoming PHP 7.4.

The version is not filled because of a variable misnaming,
which was not covered by the imported upstream patch.

To not mangle the upstream patch, add an additional patch to
fix things up.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>