apr-util: add drivers
This adds the following drivers as extra packages:

- openssl
- mysql
- odbc
- pgsql
- sqlite3
- gdbm (dbm)
- ldap

This also removes the gratuitous dependency on libsqlite3 from
libaprutil (dbd_sqlite3 was nevery packaged before, after all).

With this ldap and session-crypto support can be enabled in apache.

The mysql driver is currently marked BROKEN, because apr-util still uses
mysql server headers, which aren't part of mariadb-connector-c. But
there's work underway at apr-util upstream to fix that so that BROKEN
can be removed during a feature version bump.

The gdbm dbm driver got added together with a patch that fixes the error
handling. Patch is from Debian.

Resolves #10886

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>