apache: break-out CONFIGURE_ARGS, add layout
This gets rid of flags. For instance $(FPIC) shouldn't be forced onto
applications (see [1]).

And CONFIGURE_ARGS + CONFIGURE_VARS are broken out of Build/Configure.
This way more arguments can be added easily in the future.

The target is changed from apache to apache2 (which is used by upstream
by default). the CONFIGURE_ARGS are changed where need to enable
modules.

This also renames one patch that fixes scoreboard location (the name
004-pidfile_fix.patch didn't describe what it's doing).

Now with the OpenWrt layout in place 003-logdir_fix.patch can be
removed.

[1] https://wiki.gentoo.org/wiki/Project:Hardened/Position_Independent_Code_internals

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>