django1: Add python3-django1 package
This adds a Python 3 version of the django1 package.

This also adds a python-django1-common package that contains a
django-admin script based on the one in Debian[1]. This allows
python-django1 and python3-django1 to be installed at the same time.

python3-django conflicts with python-django1 (via python-django1-common)
and python3-django1.

This also updates older Python 3 Django plugin packages to depend on
python3-django1, and newer plugin packages to depend on "django", which
both python3-django and python3-django1 provide.

Because of this dependency on either version of Django, the MDEPENDS for
Python 3 Django plugin packages no longer functions correctly and has
been removed.

[1]: https://salsa.debian.org/python-team/modules/python-django/blob/debian/buster/debian/django-admin

Signed-off-by: Jeffery To <jeffery.to@gmail.com>