django-jsonfield: Change packaged project
Previously, this packaged "django-jsonfield" on PyPI.

The only (in-tree) package that depends on this package is
django-postoffice.

django-post-office actually depends on "jsonfield" on PyPI.[1][2]

This changes the packaged project from "django-jsonfield" to
"jsonfield".

The version packaged here is not the latest available update, but the
last version compatible with Django 1.11 / Python 2.

[1]: https://github.com/ui/django-post_office/blob/v3.2.1/setup.py#L45
[2]: https://github.com/ui/django-post_office/issues/182

Signed-off-by: Jeffery To <jeffery.to@gmail.com>