quote: disable a bunch of options
libwrap is fairly useless. It's also legacy and should not be used/

dbus is a big dependency. It's fairly useless in the OpenWrt context.

I don't know how useful netlink is.

Disabling BSD behavior results in a slightly smaller size.

Disable LDAP. No size difference but I have a feeling it pulls in
another dependency...

Signed-off-by: Rosen Penev <rosenp@gmail.com>