less: add ALTERNATIVES
This package can not be installed if you have installed less from
busybox.

Collected errors:
 * check_data_file_clashes: Package less wants to install file /bin/less
        But that file is already provided by package  * busybox
 * opkg_install_cmd: Cannot install package less.

To avoid this error, I moved it from /bin/less to /usr/bin/less.

If you install it now, it changes symlink from busybox to /usr/bin/less
/bin/less -> busybox*
/bin/less -> /usr/bin/less*

When you remove it, it changes symlink back to busybox.

That's why postint and postrm scripts are no longer needed.

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>