Revert gperf: remove
gperf is a basic tool used when building packages. It's not available on
all hosts, in which case having a host package like this one is
required. For instance the CircleCI environment does not contain gperf,
in which case building of software requiring gperf fails.

This commit reverts the gperf removal.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>