fping: install fping SUID root
fping requires either root or CAP_NET_PING to work, otherwise it is
useless.  Use INSTALL_SUID so that fping will be setuid root, and thus
it will be usable by non-root users.

fping knows to drop root priviledges after it parses the command line
and creates the ping socket.  You actually get a lot less code running
as root when you make it setuid root and run it from an unprivileged
user.

This is the same way net/iputils already handles "ping", which has the
same requirements.

Signed-off-by: Henrique de Moraes Holschuh <henrique@nic.br>