readsb: new package
This package is based on dump1090-fa with many modifications. A big
advantage over dump1090-fa is that it can connect to multiple services
like adsbexchange, piaware or fr24feed.

As the idea for this package is mostly to feed other services, the HTML
files to serve via a webserver have not been included. The service(s)
that will be fed by readsb can offer such a webpage instead.

Finally, the init script does not use local variables, as local is
undefined in POSIX.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>