sqm-scripts: Switch sch_cake dependency to new virtual package
As reported in https://github.com/openwrt/packages/issues/12072, the
imagebuilder fails due to a dependency resolution error when the userspace
packages are built using a target that has a different kernel version than
that which is being run. To resolve this, move the sqm-scripts dependency
to a new virtual package, which hopefully should be consistent with the
actual kernel module being built.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>