libssh2: switch default to OpenSSL and fix depends
All packages in the tree that depend on libssh2 _and_ have their own
dependency on a TLS library, depend on OpenSSL by default. It makes
sense for the libssh2 package to default to OpenSSL as well. This way it
does not pull in a second TLS library.

Also, the OpenSSL dependency is changed to
"+!LIBSSH2_MBEDTLS:libopenssl", which fixes build issues in case libssh2
is not selected but built anyway.

And last, Config.in file is removed and its content added to the
Makefile instead.

Closes #12108

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>