xinetd: do not save xinet.d configrations on sysupgrade
Since we can now configure xinet.d with uci, it is no longer necessary
to save the configuration in /etc/xinet.d/* when upgrading the system.

This was wrong anyway, because other packages can also store configuration
there. If this configuration get changed then the new/changed one will never get
applied, because the configration is always replaced with the saved
configration during sysupgrade.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>