python-mysqlclient: drop package
The main user for this package was Seafile.
In the meantime, Seafile switch to PyMySQL.
    https://pypi.org/project/PyMySQL/

PyMySQL seems to be a replacement for python-mysqlclient, and while it may
not be fully compatible with the MySQL API, it may be that those APIs
wouldn't be used.

This change drops this package.
If there is enough usage/reason to bring it back, we can.

For python-mysqlclient, the tag-line/description is:
```
This is a fork of MySQLdb1.

This project adds Python 3 support and bug fixes. I hope this fork is
merged back to MySQLdb1 like distribute was merged back to setuptools.
```

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>