mpd: enable libmad globally
Since making ffmpeg dependent on BUILD_PATENTED, the full variant
becomes unable to play mp3s. Change that.

libmad is superior to mpg123 because of its faster decoding speed on
soft float systems and because in MPD, it supports streams (HTTP for
example).

ffmpeg supports streams as well. Avoid libmad in that case.

Minor Makefile cleanups.

Signed-off-by: Rosen Penev <rosenp@gmail.com>