coreutils: fix compilation for when selinux is present
According to https://bugs.gentoo.org/301782 coreutils does not respect
--without-selinux correctly when libselinux.so is present on the target
system (in the staging dir).

This solves a weird issue of some coreutils programs not building (for
example stdbuf), because the configure script uses -Werror to determine
whether __attribute__((constructor)) is respected, but -Werror causes
compilation to fail because of a warning about redefinition of
HAVE_SELINUX_SELINUX_H macro.

Signed-off-by: Marek Behún <kabel@blackhole.sk>