mariadb: Use defaults and change default datadir
Use /srv/mysql as default datadir as /var/lib/mysql is in tmpfs. This
doesn't affect any existing setup as up till now it had to be always
specified in configuration. That is addressed in the second part of this
commit - init script now uses even defaults as compiled in MariaDB so
not everything has to be specified in configuration file.

Signed-off-by: Michal Hrusecky <michal.hrusecky@turris.com>