ifstat: import fixes from Debian package
Extracted from:

  http://deb.debian.org/debian/pool/main/i/ifstat/ifstat_1.1-8.1.diff.gz

Note that I also created a new git repository with these fixes:

  https://github.com/matttbe/ifstat/

The original author of these modification is:

  Goswin von Brederlow <goswin-v-b@web.de>

ChangeLog:

  * snmp.c: fix 2 pointer targets differ in signedness warnings
  * Adding upport for 64bit /proc/net/dev counters.
  * Clean up compiler warnings.

More modifications are available in the patch from the Debian project
but mostly related to the "debian" dir, man page and debug mode. Here I
only took the modifications related to the .c and .h files.

The most important fix is related to the support for 64bit counters in
/proc/net/dev instead of displaying 0 after a while.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>