rp-pppoe: increase PKG_RELEASE
In a recent commit, there was a fixed typo in config file of rp-ppoe
package. As there was no increased version in PKG_VERSION/PKG_RELEASE,
it means that fixed typo will be applied for users, who install
rp-pppoe now. Existing users will not be aware that there is an updated
package with fixed typo. They will need to do force overwrite/reinstall via opkg.

It makes a little bit complicated as we are fixing typo in conffile, but
this change will be applied to users who do not touch it. In any case,
there should be a bumped version.

Fixes: 963bdf07a3f1a6b45b507f90fb43676d600ac615 ("rp-pppoe: fix typo")
Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>