rpcd-mod-lxc: add postinst to reload rpcd on update/installation
This is dependency of luci-app-lxc and when users install that package
it is no way clear that they have to reload rpcd to get it working
correctly. Without it container listing does not work.
In general this reload should be in this package simply because other
rpcd-mod-* packages reload rpcd as well.

Signed-off-by: Karel Kočí <karel.koci@nic.cz>