elektra: explicitly disable gpgme plugin
With the addition of gpgme-config per a72d536, elektra has started
detecting the presence of gpgme and is automatically building the
plugin. Even though it's not being installed, it has broken the build on
CONFIG_ARCH="arc" due to some LFS weirdness with uClibc.

Explicitly disable it since the package in its current form doesn't
expect it to be built.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>