php7: fix build with QUILT
When building with QUILT, unlike the regular build, Build/Prepare does
not apply the patches.  So when buildconf is called with QUILT on, at
the end of Build/Prepare, it will not have the patched sources, and
build will fail.

To fix the problem, run buildconf in Build/Prepare only when QUILT is
off, and do it in Build/Configure otherwise.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>
(cherry picked from commit 89e08c188b3b0b08e936f3a1c252b6eed4e01f10)