gnunet: update to version 0.16.2
Sync GNUnet package with master branch.
This is a new major release. It breaks protocol compatibility with the
0.15.x versions. Please be aware that Git master is thus henceforth
(and has been for a while) INCOMPATIBLE with the 0.15.x GNUnet network,
and interactions between old and new peers will result in issues.
0.15.x peers will be able to communicate with Git master or 0.16.x
peers, but some services - in particular GNS - will not be compatible.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>
(cherry picked from commits
 641853c97a gnunet: update to v0.14.1
 0f4a9b6498 gnunet: update to version 0.15.3
 8863c461d4 gnunet: fix compilation with big endian
 419283ccc4 gnunet: several improvements
 0600e7400b gnunet: packaging fixes
 ba3b3454de gnunet: update to 0.16.1
 140ff2f43d gnunet: improve init script
 03309832d4 gnunet: update to 0.16.2
 6b4cb29353 gnunet: work-around sysupgrade restore missing file ownerships
)