h2o: remove useless ruby dependency
h2o is the library dnsdist uses to offer DNS over HTTPS to clients. dnsdist is the only user of h2o in this tree.

While h2o can depend on Ruby (to build mruby support), this is disabled in the OpenWRT build of h2o. Hence, the Ruby dependency is unnecessary, and removing it saves a few megabytes of disk space.

Signed-off-by: Peter van Dijk <peter.van.dijk@powerdns.com>
(cherry picked from commit ca163089384cbfe99f28fc7ea208ccb468899092)