vnstat2: add hotplug script for adding interfaces
If an interface doesn't exist yet when vnStat is started, it won't be
monitored, as only existing interfaces can be added to the database via
the vnstat command.

This adds a hotplug script which adds any configured interfaces to the
vnStat database when it goes up.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>
(cherry picked from commit 17b39cdb62eac043b72e090509723b02b0845ee0)