afuse: Add new package
This commit add new package afuse into packages feed.

afuse is an automounting file system implemented in user-space using
FUSE. It implements the most basic functionality that can be expected by
an automounter; that is it manages a directory of virtual directories.
If one of these virtual directories is accessed and is not already
automounted, afuse will attempt to mount a filesystem onto that
directory. If the mount succeeds the requested access proceeds as
normal, otherwise it will fail with an error.

Signed-off-by: Michal Hrusecky <michal@hrusecky.net>
(cherry picked from commit 01ed6858f89d94983e7bef34cd41328e3aeab594)